Refactor: [M3-6280]: SecurityQuestions
