+++ packages/manager/src/features/Profile/AuthenticationSettings/AuthenticationSettings.tsx
@@ -2,7 +2,6 @@ import * as React from 'react';
 import Divider from 'src/components/core/Divider';
 import Link from 'src/components/Link';
 import Paper from 'src/components/core/Paper';
-import SecurityQuestions from './SecurityQuestions';
 import TPAProviders from './TPAProviders';
 import TrustedDevices from './TrustedDevices';
 import Typography from 'src/components/core/Typography';
@@ -11,6 +10,7 @@ import { DocumentTitleSegment } from 'src/components/DocumentTitle';
 import { ErrorState } from 'src/components/ErrorState/ErrorState';
 import { PhoneVerification } from './PhoneVerification/PhoneVerification';
 import { ResetPassword } from './ResetPassword';
+import { SecurityQuestions } from './SecurityQuestions/SecurityQuestions';
 import { SMSMessaging } from './SMSMessaging';
 import { styled } from '@mui/material/styles';
 import { TwoFactor } from './TwoFactor/TwoFactor';
