--- packages/manager/src/features/Profile/AuthenticationSettings/SecurityQuestions/index.tsx
@@ -1,3 +1,0 @@
-import SecurityQuestions from './SecurityQuestions';
-
-export default SecurityQuestions;
